optimised the coins
reduced the triangle count on the face of the coins, seems to make the gold coin face seem a little darker, not really sure why, they should have the same smoothed normals but maybe not, either way the difference is so minimal it's hard to tell the difference. reduced triangle count from 872 triangles per coin to 827 triangles per coin, that's 45 triangles less!